frontend/backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
pandas
python-multipart
python-dotenv
